Repackage classes into packages
