+++ src/main/java/Commands/Commands.java
@@ -1,3 +1,5 @@
+package Commands;
+
 public enum Commands {
     LIST, MARK, UNMARK, DELETE, BYE, TODO, EVENT, DEADLINE;
 }
