+++ src/main/java/Commands/ListCommand.java
@@ -1,0 +1,17 @@
+package Commands;
+
+import Files.Storage;
+import Tasks.TaskList;
+import Ui.Ui;
+
+public class ListCommand extends Command {
+    @Override
+    public boolean isExit() {
+        return false;
+    }
+
+    @Override
+    public void execute(TaskList taskList, Ui ui, Storage storage) {
+        taskList.listItems();
+    }
+}
