+++ src/main/java/Commands/UnmarkTaskCommand.java
@@ -1,3 +1,10 @@
+package Commands;
+
+import Exceptions.DukeException;
+import Files.Storage;
+import Tasks.TaskList;
+import Ui.Ui;
+
 public class UnmarkTaskCommand extends Command {
 
     private String taskIndex;
@@ -8,12 +15,12 @@ public class UnmarkTaskCommand extends Command {
     }
 
     @Override
-    boolean isExit() {
+    public boolean isExit() {
         return false;
     }
 
     @Override
-    void execute(TaskList taskList, Ui ui, Storage storage) {
+    public void execute(TaskList taskList, Ui ui, Storage storage) {
         try {
             taskList.unmarkTask(taskIndex);
         } catch (DukeException e) {
