+++ src/main/java/Exceptions/DukeException.java
@@ -1,3 +1,5 @@
+package Exceptions;
+
 public class DukeException extends RuntimeException {
     public DukeException(String message, Throwable error) {
         super(message, error);
