+++ src/main/java/Exceptions/IncompleteCommandException.java
@@ -1,3 +1,7 @@
+package Exceptions;
+
+import Exceptions.DukeException;
+
 public class IncompleteCommandException extends DukeException {
     public IncompleteCommandException(String errorMessage, Throwable error) {
         super(errorMessage,error);
