+++ src/main/java/Exceptions/InvalidIndexException.java
@@ -1,3 +1,7 @@
+package Exceptions;
+
+import Exceptions.DukeException;
+
 public class InvalidIndexException extends DukeException {
     public InvalidIndexException(String errorMessage, Throwable error) {
         super(errorMessage,error);
