+++ src/main/java/Exceptions/NoSuchTaskException.java
@@ -1,3 +1,7 @@
+package Exceptions;
+
+import Exceptions.DukeException;
+
 public class NoSuchTaskException extends DukeException {
     public NoSuchTaskException(String errorMessage, Throwable error) {
         super(errorMessage,error);
