+++ src/main/java/Exceptions/UnknownCommandException.java
@@ -1,3 +1,7 @@
+package Exceptions;
+
+import Exceptions.DukeException;
+
 public class UnknownCommandException extends DukeException {
     public UnknownCommandException(String errorMessage, Throwable error) {
         super(errorMessage,error);
