--- src/main/java/ExitCommand.java
@@ -1,12 +1,0 @@
-public class ExitCommand extends Command {
-
-    @Override
-    boolean isExit() {
-        return true;
-    }
-
-    @Override
-    void execute(TaskList taskList, Ui ui, Storage storage) {
-        ui.displayByeMessage();
-    }
-}
