+++ src/main/java/Files/DukeFileWriter.java
@@ -1,3 +1,8 @@
+package Files;
+
+import Tasks.Task;
+import Tasks.TaskList;
+
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.ArrayList;
