+++ src/main/java/Parsers/Parser.java
@@ -1,2 +1,4 @@
+package Parsers;
+
 public abstract class Parser {
 }
