+++ src/main/java/Parsers/TaskInfoParser.java
@@ -1,4 +1,12 @@
-import java.util.Arrays;
+package Parsers;
+
+import Exceptions.UnknownCommandException;
+import Parsers.Parser;
+import Tasks.Deadline;
+import Tasks.Event;
+import Tasks.Task;
+import Tasks.ToDo;
+import Utils.StringUtils;
 
 public class TaskInfoParser extends Parser {
