+++ src/main/java/Tasks/Event.java
@@ -1,3 +1,10 @@
+package Tasks;
+
+import Exceptions.IncompleteCommandException;
+import Exceptions.UnknownCommandException;
+import Utils.DateTimeUtils;
+import Utils.StringUtils;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
