+++ src/main/java/Tasks/Task.java
@@ -1,3 +1,4 @@
+package Tasks;
 abstract public class Task {
     protected String description;
     protected boolean isDone;
